Change model struktur
i deleted first-, last- and secondname by register, because they arent nessesary to view or make favorites, they are in userProfil what can addet later
