+++ src/model/user.model.ts
@@ -1,57 +1,29 @@
 import { Schema, model } from "mongoose";
 import Profil from "./sensible.model.ts";
-
+import { onlyLetters as names, email, userName } from "../utils/regex.ts";
 // costum validation
 
-const names = /^[A-Za-zÄÖÜäöüßÉéÈèȨȩÑñÇç\- ]+$/;
-const email = /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
-const userName = /^[A-Za-z0-9._-]+$/;
-
 const userSchema = new Schema({
-  firstName: {
+  userName: {
     type: String,
     required: true,
     trim: true,
-    minLength: 3,
-    maxLength: 20,
+    unique: true,
     validate: {
       validator: function (val: string) {
-        return names.test(val);
+        return userName.test(val);
       },
-      message: "Firstname must contain only letters",
+      message: "Invalid username",
     },
   },
-  secondName: {
-    type: String,
-    trim: true,
-    minLength: 3,
-    maxLength: 20,
-    validate: {
-      validator: function (val: string) {
-        return names.test(val);
-      },
-      message: "Secondname must contain only letters",
-    },
-  },
-  lastName: {
-    type: String,
-    required: true,
-    trim: true,
-    minLength: 3,
-    maxLength: 20,
-    validate: {
-      validator: function (val: string) {
-        return names.test(val);
-      },
-      message: "Lastname must contain only letters",
-    },
-  },
+
   email: {
     type: String,
     required: true,
     trim: true,
     unique: true,
     lowercase: true,
+
     validate: {
       validator: function (val: string) {
         return email.test(val);
@@ -65,18 +37,6 @@ const userSchema = new Schema({
     trim: true,
     minLength: 6,
     maxLength: 60,
-  },
-  userName: {
-    type: String,
-    required: true,
-    trim: true,
-    unique: true,
-    validate: {
-      validator: function (val: string) {
-        return userName.test(val);
-      },
-      message: "Invalid username",
-    },
   },
   profil: {
     type: Schema.Types.ObjectId,
